requirements.txt
asgiref==3.2.3
bson==0.5.8
dataclasses==0.6
Django==2.2.10
django-cors-headers==3.2.0
django-jsonfield==1.4.0
django-mongoengine==0.3.3
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.1
djoser==2.0.3
mongoengine==0.19.0
mysql==0.0.2
mysqlclient==1.4.6
Pillow==7.0.0
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
